+++ algotik_tse/data/__init__.py
@@ -1,0 +1,1 @@
+"""Versioned, human-reviewable package data."""